pyproject.toml
[project]
name = "code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "bert-score>=0.3.13",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "comm==0.2.3",
    "contourpy==1.3.3",
    "cryptography==45.0.6",
    "cycler==0.12.1",
    "debugpy==1.8.16",
    "decorator==5.2.1",
    "executing==2.2.0",
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "fonttools==4.59.0",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.1",
    "google-api-python-client==2.178.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-genai==1.29.0",
    "google-generativeai==0.8.5",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.74.0",
    "grpcio-status==1.71.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "idna==3.10",
    "ipykernel==6.30.1",
    "ipython==9.4.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "kiwisolver==1.4.8",
    "lxml==6.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.5",
    "matplotlib-inline==0.1.7",
    "narwhals==2.0.1",
    "nest-asyncio==1.6.0",
    "numpy==1.26.4",
    "packaging==25.0",
    "pandas==2.3.1",
    "parso==0.8.4",
    "pdfminer-six==20250506",
    "pdfplumber==0.11.7",
    "pillow==11.3.0",
    "platformdirs==4.3.8",
    "prompt-toolkit==3.0.51",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "psutil==7.0.0",
    "pure-eval==0.2.3",
    "pyarrow==21.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydeck==0.9.1",
    "pygments==2.19.2",
    "pyparsing==3.2.3",
    "pypdfium2==4.30.0",
    "pytesseract==0.3.13",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.2.0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pywin32==311",
    "pyzmq==27.0.1",
    "referencing==0.36.2",
    "requests==2.32.4",
    "rpds-py==0.26.0",
    "rsa==4.9.1",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "spacy>=3.8.7",
    "stack-data==0.6.3",
    "streamlit==1.48.0",
    "streamlit-option-menu>=0.4.0",
    "streamlit-pdf-viewer>=0.0.26",
    "tenacity==9.1.2",
    "toml==0.10.2",
    "tornado==6.5.1",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "watchdog==6.0.0",
    "wcwidth==0.2.13",
    "websockets==15.0.1",
]
